Fix 58f196a: supervise_as is now deprecated
In the last commit I did not realize that `supervise_as` is deprecated, as
reported in https://github.com/celluloid/celluloid/wiki/DEPRECATION-WARNING

For now, let's include the deprecated module that is supposed to work until
celluloid hits 0.19.0. Eventually this gem will be removed, so we might not
need to update the code all the way through.